pyproject.toml
[tool.poetry]
name = "ocioTools"
version = "0.1.0"
description = "Package description for ocioTools"
authors = ["Suhail <suhailece@gmail.com>"]
license = "MIT"
readme = "README.md"
packages = [{include = "ociotools", from = "src"}]

[tool.poetry.dependencies]
python = "3.10.10"
opencolorio = "^2.3.2"
pyyaml = "^6.0.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
